[cmake/android] Install librtmp
After the removal of RTMP in https://github.com/xbmc/xbmc/pull/10017,
the library fails to get packaged for android because it is no more
populated in the dyload-libs variables.

As the library is still needed (https://github.com/xbmc/xbmc/pull/10079)
we manually add it to the files that we're going to package.

Fixes:
cp: cannot stat 'kodi-build-android/install/lib/librtmp.so': No such file or directory
Makefile:121: recipe for target 'libs' failed