[cmake] Introduce ENABLE_STATIC_LIBS option
Add an option ENABLE_STATIC_LIBS that defines if the buildsystem
creates one static library per subdirectory in xbmc/ or compiles
everything into one big library.

The option defaults to true, except for Visual Studio and Xcode,
because these generators work much better with a single big library.
The project structure in Visual Studio and Xcode is closer to the
filesystem, which cannot be realized otherwise, because they show
separate projects for each subdirectory.

Several related changes are introduced in this commit:
- dependency management had to be adapted to work in both cases
- include directories are propagated through
  target_include_directories(... INTERFACE ...)
- dvdnav and crossguid are made a project wide dependency, because they
  are widely used within the project.