Default keymaps for WeTek USB remotes
Successor to https://github.com/xbmc/xbmc/pull/8290

This is using updated and safe keymaps that work on all three of the current WeTek USB remotes. This has gotten really good feedback from users and from people on the LibreELEC team.

This does not currently impact using these remotes in Android, so this is linux only (LibreELEC/OpenELEC, OSMC, etc).