[cmake] Add a 'destination directory' parameter to copy_file_to_buildtree
The function copy_file_to_buildtree is used to mirror files to the
build directory, so that Kodi can be launched from there.

The previous implementation of that function only allowed to mirror
files to the same tree structure. With this change files can be
mirrored also into different directories as this is needed for some
use cases (DLLs on Windows, certificates on IOS/OSX).

This changes the following parts:
- Patterns specified in installdata now support an optional
  second parameter that specifies where the files should be mirrored to.
  This parameter is optional. If specified it points to the destination
  directory (relative to CMAKE_BINARY_DIR) otherwise the files are
  mirrored in the same tree structure under CMAKE_BINARY_DIR.
- Mirror files for in-source-builds where necessary.
- Remove the relative parameter from copy_file_to_buildtree.
  Its API was confusing and it's not necessary because the files are
  always relative to CORE_SOURCE_DIR.
- Remove Windows specific hook to mirror additional files and
  use the newly introduced mechanism instead.