Fix crash because of heap mismatch
When running a debug version of Kodi we allocate
plugin_path on the debug heap and then pass it to
cpluff to free. cpluff being a release version does not
use the debug heap and crashes because it tries
to free an invalid address.

Also added some minor safety to the string copy