[cmake] Fix options
Commit fe50635fb7a09d1576890a5d548429d55af00221 changed the behavior
of optional dependencies. ON now makes them a required dependency,
AUTO an optional one and OFF disables it. Options that were not listed
before were implicitly disabled (and supposed to be overridden in
platform specific files).

This change removes explicit options for optional dependencies from
CMakeLists.txt because they were now incorrectly made required
dependencies. In addition it adds options that default to off, because
they are now optional dependencies.