peripherals: Log when device's add-on doesn't provide button maps
The button map allows Kodi to translate driver input (anonymous button, hats
and axes) into the button format of the requested controller type.

When a device is connected, Kodi first checks if the device came from an
add-on. If so, Kodi checks if the add-on provides button maps for its
device. Otherwise, Kodi looks for ANY add-on that can map buttons.