[pvr] never play recordings minimized
This is part of the "remove unused parameters" warnings. I think the setting for this was originally aimed at live TV only (where you control if playback starts in the preview window or if it goes directly to fullscreen), it thus makes no sense for recordings which are just like regular files in that sense.