redesign texture loading to avoid locking the texture bundle during skin reloads.
Cuts down on indirection and caching, simplifies design

Not entirely sure if this will negatively impact performance in a
noticeable way. Since we do open and close the texture bundle a few
times it's likely a bit slower in practice but it should only be
during boot and skin reload.

Don't initialize the xbtfreader on every call to open

Merge CXBTFBase and CXBTReader as there's no other classes inheriting from
base
Remove some unused methods