[binary-addons] Rename DEPENDS_PATH into ADDON_DEPENDS_PATH
DEPENDS_PATH is used twice in the code build system it's set to the
system depends path (via toolchain file). In the addon system it's
used to point to the place where addon dependencies should be built.

This change renames the variable for binary-addons to
ADDON_DEPENDS_PATH. The toolchain file can then be used for local
addon builds.