[VideoPlayer/Process] Make PLATFORM_OVERRIDE define unique
The way our CMake build system works, we cannot specify defines on a
per file level. Mainly because of the ENABLE_STATIC_LIBS option that
is necessary to create nice Xcode/VS projects.

This commit changes the PLATFORM_OVERRIDE define in VideoPlayer to
a project unique PLATFORM_OVERRIDE_VP_PROCESSINFO.