[cmake] Add add_platform_override function
A generic header includes a generated file. This generated file is
either empty (no override for current platform), or contains an include
line to the platform specific header.

This gives IDEs the chance to show the correct platform override header
from the source dir (no copying).

CMake is only retriggered if a platform to override is added or removed
so we avoid unnecessary reconfigure or even make runs.