MMALCodec: Take more care with the output lock mutex
There were reports of "GetPicture called but m_output_ready is empty"
error in log.

It's been hard to reproduce but the only way that seems possible is for
m_output_ready.size to produce an inconsistent result between calls
to CMMALVideo::Decode and CMMALVideo::GetPicture.

Looking closely we don't always have the m_output_mutex lock when
testing the size, so do it more carefully.