[cmake] Properly find correct libcec version
Just setting a pkgconfig minimum version doesn't work and just causes
pkgconfig to not find an older package. But since our CMake modules
don't trust package config and validate that the include directory
really exists, it would still just find whatever version is installed.

The proper way is to either call find_package with VERSION or set
<Module Name>_FIND_VERSION inside the module.

Additionally: If we don't have a version from package config, find it
in the header file.