Rework SQL used by GetFilter for artists, albums and songs nodes.
Consistent passing of filter when drilling down to albums or songs for a selected artist. Change build of "albums" and "songs" type node query where clauses in GetFilter to use role and albumartistsonly flag combinations.

For the default node (role = 1) when albumartistsonly only show albums, and the songs on those albums, by that artist, exclude those albums that have songs they are a song artist for but are not credited with the overall album. For other specific roles (custom nodes) when albumartistsonly only show albums with songs that have that artist in that role AND as album artist. When looking at all contributors (role < 0) albumartistsonly has no meaning, and both song_artist and album_artist tables are examined.

Without this change:
1) drilling down to albums for an album artist gives any album with a song credited to them, including both compilation albums and where they have collaborated on a song on an album by someone else.
2) drilling down to albums with songs by a composer (or other role) or the songs themselves, also gives any album or songs where they are the album artist.

Apply genre and artist options for "albums" node correctly together using subquery on song. Without this it was selecting albums that had song(s) by the artist and song(s) with that genre, but not ensuring they were the same song.
Redesign artists SQL