ListItems: Rename "emulator" property to "gameclient"
This change apparently got dropped in a RetroPlayer rebase. The reason is
because gameclient is more generic than emulator, as the game might be a
launchable asset instead of a ROM. Also, I would expect "emulator" to be
a human-readable string, e.g. "PCSX Re-armed", whereas "gameclient" is
more obviously an add-on ID.