Move game extensions to advanced settings
This imports the 97 extensions used by our 58 libretro cores to advanced
settings.

This fixes a UX problem where no games show up in MyGames if no add-ons are
found locally or in a remote repo. It also saves some add-on DB hits,
improving performance in areas that require game extensions.