[addons] move save of settings from CAddonDll::Stop() to CAddonDll::Destroy()
The stop becomes removed on next commits. Reason about 'Stop' removal is a
fissing start and also equal on add-on to destroy. Has made sense so long
a addon has only be supported one instance to prevent a standard function
on every type. But with use of multiple types makes it no sense.