[addons] add new versions check function
This alow to remove on every add-on type the same version check functions
and mainly remove the add to add-ons, becomes now done only by the headers.

versions.h is the first final part of add-on system rework. With them are
all versions set to only one space who makes it more easy for changes and
reviews.