[addons][cmake] remove the create of addon.xml for guilib
This removes the generation of addon.xml for guilib. There are only two ways, one
to add the create to every addon.xml, who makes a lot of cmake parts who makes
it hard to add new parts or changes.

The other for me better way is to remove the create and do it by hand like the
others.