[addons] add first initial new addon to kodi callbacks
This add the first callbacks over the new way to add-on header. With them
is on add-on start the table already set and something like "XBMC->RegisterMe(hdl)"
no more needed.

There comes a bigger cleanup of "AddonBase.h" after them is in.

Currently don't touch this the old way and all old style add-ons don't need a
rework.