[addons] change screensaver classes to the final style
This changes the screensaver classes to the way used on final system.
The only part who not complete correct is the get of addon class and
start of the Python Screensaver.

About the Python Screensaver, after this changes it is startet in a better
way and no more the start screensaver class needed.