[addons] change 'AddonsFromRepoXML' to new way over properties
This is a major rework part where for repository reads for add-on's no
more the related CAddon or CAddonDll becomes created, what was never
needed there.

This change bring currently a compile fault, as long the other parts
not changed to the new way, only separated here as alone for better
review.