[addons] remove the use of "FromExtension" on CPeripheralAddon
This type use now the new (not from cpluff) generated CAddonInfo and can
read extension values always. For this are the set of related values moved
to constructor and the "FromExtension" is removed.

Also include this commit changes to "CAddonExtensions" who allows read of
values with right return type (e.g. string, int or bool)