Inject a channel EPG creation step into CPVRTimersUpdateJob::DoWork()
The problem I'm trying to solve is that when new channels are added by
the backend via TriggerChannelUpdate() after the PVR Manager has run the
initial CPVREpgsCreateJob there is no call to create the EPG for the
newly added channel(s), necessitating a shutdown and restart of the PVR
Manager.  This change represents the point where I was able to inject
the creation step successfully with the smallest impact to the PVR code.