Controller dialog: Update button label
The previous label, "Fix skipping", is no longer relevant once the changes
in PR 11571 are in, because skipping occurs much more rarely.

"Ignore input" is a better title, because it reflects what is actually does,
and should be less confusing for users who need input ignored but don't
necessarily experience skipping.