Fix errors from axes discovered during mapping
On some platforms, such as OSX, an axis is not discovered until it is first
moved. This causes problems with button mapping because we can't detect the
axis's center by the initial position.

To fix this, the user is shown a dialog that prompts them to press all
analog buttons, guaranteeing they'll have an initial value next time.