[addons] add multiple instance support to Kodi
This add the support to handle multiple instances together on one
add-on.

A addon class stays so long created as it is used by instances,
after the last instance release it becomes the addon class destroyed.

Further stays the data on CAddonDll and if them call for a type an
"CreateInstance" on add-on becomes also the first time a "Create"
performed, every other "CreateInstance" call do know that the
"Create" is already done and becomes skipped.