[addons] make inputstream check static
This change the check for supported addon from inputstream to
static. This is done to prevent a class creation and to
bring ADDON::CInputStream and CInputStreamAddon together (later).

It should be prevent on the static function to load the dll and to
use data only what available on addon.xml.