Get rid of our emulated GetDiskFreeSpaceEx method
Decided to unify some of our low level file handling
methods on the std::filesystem idea. As it's not standardized
yet and implementations are experimental this is a simple
version that suits our current needs. The idea is that once platforms
start offering non-experimental versions of std::filesytem we can
easily move over to using them.