RenderManager: Force frames to output when reconfiguring renderer
Currently the clock may be paused before the final queued frame is rendered.
This prevents the queue of pictures to render from being drained and
so stops the PRESENT_IDLE state being reached.

This results in a 5 seconds pause and the following warning:
"CRenderManager::Configure - timeout waiting for state"

This has been observed on Pi in DVD menus.