Remove '[CR]', formatting should be left to skins
These [CR] kind of dictate layout/formatting, this should be left as plain as possible for skins to manipulate as needed [imho] :)
Usage scenario: I want to place the text provided by kodi on one 'label' line, but the [CR] hardcoded means it has to be a 'textbox' or string of same height to match height of various linesnow forced by '[CR]', ie. if one straight line of text I can scroll it, I don't have to make a large box.
Thanks.