Update AndroidKey.cpp
Extending support for the physical QWERTY keyboard in Android Kodi with processing of AKEYCODE_MOVE_HOME and AKEYCODE_MOVE_END key codes. Currently, these keyboard events are just ignored. In large Kodi list views jumping on FirstPage and LastPage is very convenient. I do not see any reason why it should not be supported in the Android Kodi , as it is on other OS platforms.