[pvr] change to new addon to kodi callback way
This is the first major change to become multiple instances to work.

There are the PVR callbacks moved to the CPVRClient class. Reason is
to have this parts together and later with multiple instances to call
always right related CPVRClient class.

On addon is currently nothing needed to change, only a recompile to
match them.