[inputstream] change to new addon to kodi callback way
This is the first major change to become multiple instances to work.

There are the callbacks moved to the ADDON::CInputStream class. Reason is
to have this parts together and later with multiple instances to call
always right related ADDON::CInputStream class.

On addon is currently nothing needed to change, only a recompile to
match them.