[screensaver] change addon interface to the coming style
This request change screensaver to the complete C++ style of
related interface headers.

This style is also for multiple instances used, but not needed
on Screensaver. This request is mostly to see the new style.

The other thing is that the kodi-agile and master are equal
on addon side for screensavers.