[addons] add addon audio engine callbacks on new way
This add the currently with libKODI_audioengine.h done part to the
new style.

Due to missing addons who use them was a complete test currently not
possible and must be check in future, from the base is the work
equal to the old.

The old parts becomes in next commit removed.