Expand tempo range to 0.5x-5.0x
This expands the range of tempo settings available via tempoup /
tempodown commands from 0.75x-1.5x to 0.5x-5.0x.

There is a minor chance that this will interact with fast forward
behavior, but that seems unlikely to be problematic: if the user
modifies the tempo at 2x or 4x fast forward, the new tempo setting
will take effect, and will be used as the base when multiplied
or divided by 2 if the user then presses "rewind" or "forward".

When the value exceeds 32x or gets down below 1.55x, the "rewind"
and "forward" commands will treat it as 1, escaping any odd
behavior. If necessary, we can clamp values in "rewind" and
"forward" to the nearest power of 2, but that seems like overkill
in this situation.