[addons] add addon filesystem callback functions to new style
This is the next step to add addon callbacks to the new way.
After them are all parts who already in old libXBMC_addon.h included on
new interface headers.

As next comes the change of GUI callbacks.

Because of the size is the doxygen part not included here and comes in another
commit.