[addons] add new CAddonInfo class
This add the class who becomes extended step by step to replace cpluff.

Currently include it only the static functions and table who was before
on other place.

This is done on this commit and request to hold coming request smaller
and prevent parts who not direct related to changes.

As next requests comes the basic read of addon.xml and to go over there.