[addons] add first gui control addon class for buttons
This add the first class on new addon interface to edit a button
control.

Hold on this alone to see better how the controls becomes used
on new addon system.

On next request are all controls together.

Here the smallest example to edit a control on new way:
```cpp
// this = The addon window class
// 33 = The on skin.xml used control id
kodi::gui::controls::CButton(this, 33).SetVisible(false);
```