[addons] move AddonProps to new CAddonInfo class
This add all previous on AddonProps defined values to
CAddonInfo who was added by commits before.

On this change is mostl only an all parts a rename done.

The commit comes in alone to hold next changes smaller
and better for review.

The CAddonInfo becomes on comming changes more and more
changed to allow multiple addon instances.

For this is it later not a static definition in CAddon
and also replace the old cpluff system.