Add an option to pass the ffmpeg cookies http protocol avoption instead of sticking cookie values inside the ffmpeg headers http protocol avoption.
This is preferable for streams that periodically update authentication values (such as a session token) that are stored as cookies. If the cookie values are specified in the headers avoption, it conflicts with the http protocols cookies option, which does get updated properly.

The format is different. The value used with the headers avoption is just the cookie value, whereas, the cookies avoption expects a newline delimited list of Set-Cookie values (eg: everything after ‘Set-Cookie: ‘).