[addons] change visualization to new way
This commit change the visualization addon to the new C++
based addon style.

Also is the visualization process optimized and much more
safe (found during test of old way a lot of faults and crashes).

For this need every visualization addon a update to new style.

Further are parts removed who was never from Kodi and Addon used.