Add generic version of StringUtils::Split
When the target container is not an std::vector (e.g. an std::set), the
result of Split() must be copied because always returns a std::vector.
Therefore, add a generic version StringUtils::SplitTo that splits a
string into an output iterator irrespective of the underlying container.

The existing Split() functions are kept for compatibility and
reimplemented using SplitTo().