[addons] change CAddonInfo usage to a shared_ptr
This change the on CAddon inserted CAddonInfo who was before static
included as shared_ptr.

To have as shared_ptr is a major requirement to allow several types
on one addon together.

There becomes be on next steps for e.g. CPVRClient the same CAddonInfo given
as for maybe ADDON::CInputStream if on same addon.

Further becomes in future (other requests) the CAddonInfo with this
shared_ptr stored on CAddonManager in a list.