[inputstream] change addon interface to new C++ style
This commit change the inputstream addon system to the new C++ style
and remove the old parts.

The parts from here are taken from kodi-agile and becomes mainly improved
and optimized from @peak3d on next changes.

Currently it supports only the same as the old, only a new interface way.