fix massive dead loop by use of "CUtil::MakeShortenPath"
To see fault must be somewhere from settings a folder selected where
it start with "/../" to reduce length.

This freeze complete Kodi and can only be closed by a force,
is also confirmed fault on libreelec (tested on raspberry).

There becomes std::string::npos required to exit loop who can never come!
This is replaced to a "0" who works.