Use new COSScreenSaverManager in Application
This changes the rules and functionality of the OS screen saver
inhibition to be more in line in with how modern desktop environments
work. Constant resetting of the screen saver is not necessary usually.
If it is, the platform-specific backend will have to implement it.
On desktop machines (that have an OS screen saver), the default should
be to use that. If the user then chooses to activate the Kodi screen
saver, this should completely override the OS screen saver as far as
permitted by the desktop environment. During video playback etc.,
the OS screen saver must always be inhibited regardless of whether
the Kodi screen saver is enabled.
In embedded environments (that do not have an OS screen saver), the
default should be to use the Kodi screen saver.