[peripheral] change addon interface to new C++ style
This change also the peripheral addon system to use C++ class as
interface to Kodi.

There is only the interface itself changed, the functions style is still
the same as before with the use of PeripheralUtils.h (kodi_peripheral_utils.hpp).
My experience for the system is not good enough to change also the list functions
to use direct a vector on addon (not sure how I can test every function).

The test of them here was good and has played some games :-) .