[addons] make doxygen for c++ addons working
This change make the doxygen parts for binary addons working and usable.

There is still a lot needed to become it final, but the currently available
addon callbacks can be now found on a html file.

This helps also me a lot to find fast a function and his values :-)