Change CVideoSync stop flag from atomic bool to CEvent
Some windowing implementations do not need an extra thread to update
the video clock. If they just get a bool to know when to stop the
thread (that does nothing), they have to wake up every few milliseconds
to poll the flag. Each wake-up costs a context switch that is not
needed.
Passing a stop CEvent instead is more flexible: CVideoSync implementers
can still poll the flag if needed, but it is also possible to let
the whole thread sleep until the stop is requested without intermittent
wake-ups.