RetroPlayer: Fix black screen during game
The RenderManager (or somewhere in the rendering pipeline) no longer checks
for DVD_NOPTS_VALUE, so passing a pts with this value caused RenderManager
to think the frame is far in the future.

By setting it to zero, RenderManager should never add any extra delays
waiting for pts to arrive.