Fix handling of touch gestures that are in progress when switching windows
If windows are switched while a touch gesture is active, the old window
would not get the GESTURE_END event and thus e.g. not release its
exclusive mouse control state, while the new window gets gesture events
for a gesture it does not even know has started.

So when switching windows, inform the old window that the touch gesture
has been aborted, and do not relay any gesture events to the new
window until a new gesture starts.

Signed-off-by: Philipp Kerling <pkerling@casix.org>