VDPAU: fix potential crash
This patch fixes a potential bug that may cause a crash.
If "XOpenDisplay" failed to open XServer, "dl_vdp_device_create_x11" would cause a crash since "m_display" is null.
I think it's unsafe to assume that the library functions would be correct. It would be better if we could handle the errors properly.