Add events and functions allowing resolution changes from windowing
Depending on the windowing implementation, there might be cases
when the Kodi resolution must be updated not because of a request
of the application, but because the windowing system demands it.
In this case, it does not make sense to call e.g. SetFullScreen
on windowing because windowing is the entity that initiated the
resolution change in the first place.
To simplify control flow in windowing implementations that need
this functionality and make it more explicit what is going on,
add new modesetting event that windowing can generate and
then get informed when the resolution was switched. Similarly,
call back into windowing when an XBMC_VIDEORESIZE event is
processed.