[addons] fix addon install if not present before
On last changes comes the problem that a addon install who was not present
before has not worked.

This comes from a try to disable a addon who is not present and brought a
false.

With them becomes before checked that a addon is present, otherwise is this
step ignored.