Joysticks: Fix accelerometers preventing screensaver
This fixes accelerometers on PS4 controllers preventing the screensaver by
only monitoring mapped buttons and axes. By changing from IDriverHandler
to IInputHandler, we receive input for higher-level controller features
instead of lower-level driver events. If the driver input is not mapped
to a controller, the IInputHandler will receive no input, meaning it
can't interrupt a screensaver.