RetroPlayer: Cache/pre-compile output shaders
Prevents stutering induced from the compilation of output shaders, when switching between them in the Video Filters OSD.

Commits by garbear:

Revert some logic

The reason we don't test renderers is because they might not be created
yet. Buffer pools, on the other hand, are created when RPRenderManager is
created, so we ask them about support. Buffer pools, in turn, query either
static functions of the renderer, or process info (which is also created
upon startup), or local properties of the buffer pool.

Revert some added logic

The default scaling method must be supported by at least one renderer, so
we don't need to check support here.

Revert changes to GUI texture renderer

Scaling method support is implicitly defined by the types of buffer pools
created. For example, on Windows, only linear scaling is supported, so
only a buffer pool that supports linear scaling is created.

Simplify shader construction

The lambda approach succeeds in reducing the amount of code, but adds
complexity. In this case, because it's only a few lines, it would be
better to simply duplicate the code to make the code much more
understandable. We're not writing code for the computer, we're writing
code for people. We want the code to be as straight-forward as possible,
and this goal trumps saving a few lines.

That comment withstanding, we could move handling of the default scaling
method to another place and greatly simplify this function.

Move output shaders to buffer pool

squashme