Add Information Provider dialog for selection of artist and album scraper addons, and their settings.
Improve add music source work flow by making setting default music info providers available from a "settings" button on the "Add to library" yes/no prompt.

Improve the UI for the Change Info Provider facility for artists and albums.
Use a single scraper view of this dialog from the context menu on artists and albums nodes. This replaces using the SetContent dialog and a sub-menu in GUIWindowMusicNav with use of list of options on how to apply settings in Info Provider dialog itself.