[Android] Remove button ID overloading
Button presses are sent for multiple IDs. Previously, button presses were
multiplexed to the same button ID.

However, the upper layers are able to handle this. Button presses that
follow immediately after another (such as when a press generates an event
that is sent for two button IDs) are ignored in the controller configuration
dialog. So only the first button ID is mapped. Unmapped presses are then
ignored.