[gui][fix] need splash window in history to render it properly
This will revert the WINDOW_SPLASH condition in CGUIWindowManager::AddToWindowHistory introduced by 9f6eb4d3dfe35bec36f50e69a3fd81058e4334af because we need the splash in history to render it properly. Without GetActiveWindow never return the CGUIWindowSplash and therefore it's not rendered. We remove the splash window right after the GUI is ready.