Game API v1.0.34: Keyboard mapping support
This routes keyboard input through the peripehals system, which uses an
add-on to translate raw key presses into higher-level controller features.

To differentiate between raw key presses and key presses identified by
controller features, the raw key-handler`IKeyboardHandler` interface was
renamed to `IKeyboardDriverHandler` and `IKeyboardInputHandler` was added
to handle input for controller features.

The main functional change here is switching CGameClientKeyboard from a
driver handler to an input handler, which allows the keys to be remapped
with a peripheral's buttonmap.