guilib/GUIBaseContainer: use std::list for CGUIListItemLayout
This reduces the overhead, because resizing the std::vector will
create lots of temporary CGUIListItemLayout copies.  A std::list never
needs to move its items.

These lists are very small, so the overhead for iterating it is
negligible, but copying a CGUIListItemLayout (which is not movable) is
very expensive because a rather large tree of objects needs to be
copied.