[AudioEngine] add cleanup for alsa sink
This adds and registers a static cleanup method for ALSA sinks.
It is needed to stop the static device monitors that are specific to ALSA.

This fixes a crash when kodi is stopped - race condition caused by
the hctl monitor that fires a device changed event when windowing is
already down.