Move CStereoscopicsManager init/deinit to CServiceManager
This part is tricky. Identify the closest stage where the instance is
initialized and move it to CServiceManager.

In this case, it was pretty easy. Init occurred right before Stage 3, so we
move it into CServiceManager at the top of Stage 3.

Notice how deinit is done at the same relative place in the opposite order.

Before, CStereoscopicsManager was never deinitialized, so adding a deinit
step in CServiceManager might cause crashes on shutdown. Regardless, we
must be strict in CServiceManager, so we'll have to deal with shutdown
crashes if any arise.